Рефакторинг функций отрисовки/удаления компонентов
