+++ src/constants.js
@@ -32,8 +32,3 @@ const COUNTRIES = [
 ];
 
 export {DOOMY_TEXT, NAMES, COUNTRIES};
-
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
